pyproject.toml
[project]
name = "flashcard-tgbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysqlclient==2.2.7",
    "pre-commit>=4.5.0",
    "python-decouple==3.8",
    "python-telegram-bot==22.5",
    "sqlalchemy==2.0.44",
]
